FIX benchopt prepare correctly parse parameters
