+++ benchopt/tests/test_runner.py
@@ -1,13 +1,8 @@
 import pytest
 import inspect
 
-from benchopt import BaseDataset
 from benchopt.cli.main import run
 from benchopt.results import read_results
-from benchopt.benchmark import _check_patterns
-from benchopt.benchmark import _extract_options
-from benchopt.benchmark import _extract_parameters
-from benchopt.benchmark import _list_parametrized_classes
 from benchopt.utils.temp_benchmark import temp_benchmark
 from benchopt.tests.utils import CaptureCmdOutput
 
@@ -62,211 +57,6 @@ def test_skip_api(n_jobs):
 
     out.check_output(r"test-solver\[should_skip=False\]: done", repetition=1)
     out.check_output("Solver#RUN", repetition=1)
-
-
-def _assert_parameters_equal(instance, parameters):
-    for key, val in parameters.items():
-        assert getattr(instance, key) == val
-
-
-class _DatasetTwoParams(BaseDataset):
-    """Used to test the selection of datasets by keyword parameters."""
-    name = "Test-Dataset"
-    parameters = {'n_samples': [10, 11], 'n_features': [20, 21]}
-    def get_data(self): pass
-
-
-def test_filter_classes_two_parameters():
-    # Test the selection of dataset with optional parameters.
-
-    def filt_(filters):
-        return list(_list_parametrized_classes(*_check_patterns(
-            [_DatasetTwoParams], filters
-        )))
-
-    # no selection (default grid)
-    results = filt_(["Test-Dataset"])
-    assert len(results) == 4
-    _assert_parameters_equal(results[0][0], dict(n_samples=10, n_features=20))
-    _assert_parameters_equal(results[1][0], dict(n_samples=10, n_features=21))
-    _assert_parameters_equal(results[2][0], dict(n_samples=11, n_features=20))
-    _assert_parameters_equal(results[3][0], dict(n_samples=11, n_features=21))
-
-    # select one parameter (n_samples)
-    results = filt_(["Test-Dataset[n_samples=42]"])
-    assert len(results) == 2
-    _assert_parameters_equal(results[0][0], dict(n_samples=42, n_features=20))
-    _assert_parameters_equal(results[1][0], dict(n_samples=42, n_features=21))
-
-    # select one parameter (n_features)
-    results = filt_(["Test-Dataset[n_features=42]"])
-    assert len(results) == 2
-    _assert_parameters_equal(results[0][0], dict(n_samples=10, n_features=42))
-    _assert_parameters_equal(results[1][0], dict(n_samples=11, n_features=42))
-
-    # select two parameters (n_samples, n_features)
-    results = filt_(["Test-Dataset[n_samples=41, n_features=42]"])
-    assert len(results) == 1
-    _assert_parameters_equal(results[0][0], dict(n_samples=41, n_features=42))
-
-    # get grid over one parameter (n_samples)
-    results = filt_(["Test-Dataset[n_samples=[41,42], n_features=19]"])
-    assert len(results) == 2
-    _assert_parameters_equal(results[0][0], dict(n_samples=41, n_features=19))
-    _assert_parameters_equal(results[1][0], dict(n_samples=42, n_features=19))
-
-    # get grid over two parameter (n_samples, n_features)
-    results = filt_(["Test-Dataset[n_samples=[41,42], n_features=[19, 20]]"])
-    assert len(results) == 4
-    _assert_parameters_equal(results[0][0], dict(n_samples=41, n_features=19))
-    _assert_parameters_equal(results[1][0], dict(n_samples=41, n_features=20))
-    _assert_parameters_equal(results[2][0], dict(n_samples=42, n_features=19))
-    _assert_parameters_equal(results[3][0], dict(n_samples=42, n_features=20))
-
-    results = filt_(
-        ["Test-Dataset[n_samples=[foo,bar], n_features=[True, False]]"])
-    assert len(results) == 4
-    _assert_parameters_equal(
-        results[0][0], dict(n_samples="foo", n_features=True)
-    )
-    _assert_parameters_equal(
-        results[1][0], dict(n_samples="foo", n_features=False)
-    )
-    _assert_parameters_equal(
-        results[2][0], dict(n_samples="bar", n_features=True)
-    )
-    _assert_parameters_equal(
-        results[3][0], dict(n_samples="bar", n_features=False)
-    )
-    # get list of tuples
-    results = filt_(
-        ["Test-Dataset['n_samples, n_features'=[(41, 19), (42, 20)]]"])
-    assert len(results) == 2
-    _assert_parameters_equal(results[0][0], dict(n_samples=41, n_features=19))
-    _assert_parameters_equal(results[1][0], dict(n_samples=42, n_features=20))
-
-    # invalid use of positional parameters
-    with pytest.raises(ValueError, match="Ambiguous positional parameter"):
-        filt_(["Test-Dataset[41,42]"])
-
-    # invalid use of unknown parameter
-    with pytest.raises(ValueError, match="Unknown parameter 'n_targets'"):
-        filt_(["Test-Dataset[n_targets=42]"])
-
-
-class _DatasetOneParam(BaseDataset):
-    """Used to test the selection of dataset with a positional parameter."""
-    name = "Test-Dataset"
-    parameters = {'n_samples': [10, 11]}
-    def get_data(self): pass
-
-
-def test_filter_classes_one_param():
-    # Test the selection of dataset with only one parameter.
-
-    def filt_(filters):
-        return list(_list_parametrized_classes(*_check_patterns(
-            [_DatasetOneParam], filters
-        )))
-
-    # test positional parameter
-    results = filt_(["Test-Dataset[42]"])
-    assert len(results) == 1
-    _assert_parameters_equal(results[0][0], dict(n_samples=42))
-
-    # test grid of positional parameter
-    results = filt_(["Test-Dataset[41,42]"])
-    assert len(results) == 2
-    _assert_parameters_equal(results[0][0], dict(n_samples=41))
-    _assert_parameters_equal(results[1][0], dict(n_samples=42))
-
-    results = filt_(["Test-Dataset[foo, 'bar', None]"])
-    assert len(results) == 3
-    _assert_parameters_equal(results[0][0], dict(n_samples="foo"))
-    _assert_parameters_equal(results[1][0], dict(n_samples="bar"))
-    _assert_parameters_equal(results[2][0], dict(n_samples=None))
-
-    # test grid of keyword parameter
-    results = filt_(["Test-Dataset[n_samples=[foo,True]]"])
-    assert len(results) == 2
-    _assert_parameters_equal(results[0][0], dict(n_samples="foo"))
-    _assert_parameters_equal(results[1][0], dict(n_samples=True))
-
-    # invalid use of both positional and keyword parameter
-    with pytest.raises(ValueError, match="Invalid name"):
-        filt_(["Test-Dataset[41, n_samples=42]"])
-
-
-def test_extract_options():
-    basename, args, kwargs = _extract_options("Dataset")
-    assert basename == "Dataset"
-    assert len(args) == 0
-    assert kwargs == dict()
-
-    basename, args, kwargs = _extract_options("Test-Dataset[n_samples=41]")
-    assert basename == "Test-Dataset"
-    assert len(args) == 0
-    assert kwargs == dict(n_samples=41)
-
-    basename, args, kwargs = _extract_options("n_samples[n_samples=41,]")
-    assert basename == "n_samples"
-    assert len(args) == 0
-    assert kwargs == dict(n_samples=41)
-
-    basename, args, kwargs = _extract_options(
-        "Dataset[n_samples=[41, 42], n_features=123.0]")
-    assert basename == "Dataset"
-    assert len(args) == 0
-    assert kwargs == dict(n_samples=[41, 42], n_features=123.)
-
-    with pytest.raises(ValueError, match="Invalid name"):
-        _extract_options("Dataset[n_samples=42")  # missing "]"
-    with pytest.raises(ValueError, match="Invalid name"):
-        _extract_options("Dataset[n_samples=[41, 42]")  # missing "]"
-
-
-def test_extract_parameters():
-    # Test the conversion of parameters.
-
-    # Convert to a list
-    assert _extract_parameters("42") == [42]
-    assert _extract_parameters("True") == [True]
-    assert _extract_parameters("foo") == ["foo"]
-    assert _extract_parameters("foo-bar") == ["foo-bar"]
-    assert _extract_parameters("foo.bar") == ["foo.bar"]
-    assert _extract_parameters("1, 2, 3") == [1, 2, 3]
-    assert _extract_parameters("1e1, -.1e-3, 12.e+1") == [10.0, -0.0001, 120.0]
-    assert _extract_parameters("foo42e1, bar1.0e4") == ["foo42e1", "bar1.0e4"]
-    assert _extract_parameters("foo_4e2, bar01e3") == ["foo_4e2", "bar01e3"]
-
-    assert _extract_parameters("42, True, foo") == [42, True, "foo"]
-    assert _extract_parameters("42, (1, 2, 3)") == [42, (1, 2, 3)]
-    assert _extract_parameters("foo,bar ") == ["foo", "bar"]
-    assert _extract_parameters("foo, bar") == ["foo", "bar"]
-    assert _extract_parameters("foo,bar,") == ["foo", "bar"]
-    assert _extract_parameters("'foo, bar'") == ["foo, bar"]
-    assert _extract_parameters('"foo, bar"') == ["foo, bar"]
-    assert _extract_parameters("foo, (bar, baz)") == ["foo", ("bar", "baz")]
-
-    # Convert to a dict
-    assert _extract_parameters("foo=(bar, baz)") == {'foo': ('bar', 'baz')}
-    assert _extract_parameters("foo=(0, 1),bar=2") == {'foo': (0, 1), 'bar': 2}
-    assert _extract_parameters("foo=[100, 200]") == {'foo': [100, 200]}
-    assert _extract_parameters("foo=/path/to/my-file,bar=baz") == \
-        {'foo': '/path/to/my-file', 'bar': 'baz'}
-    assert _extract_parameters("foo=[\\path\\to\\file,other\\path]") == \
-        {'foo': ['\\path\\to\\file', 'other\\path']}
-
-    # Special case with a list of tuple parameters
-    assert _extract_parameters("'foo, bar'=[(0, 1),(1, 0)]") == \
-        {'foo, bar': [(0, 1), (1, 0)]}
-    assert _extract_parameters('"foo, bar"=[(0, 1),(1, 0)]') == \
-        {'foo, bar': [(0, 1), (1, 0)]}
-
-    for token in [True, False, None]:  # python tokens
-        assert _extract_parameters(f"{token}") == [token]
-        assert _extract_parameters(f"'{token}'") == [token]
-        assert _extract_parameters(f"\"{token}\"") == [token]
 
 
 # Under windows, the function needs to be pickleable
